pyproject.toml
[project]
name = "github-stars-search"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.3.0",
    "faiss-cpu>=1.12.0",
    "numpy>=2.3.4",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.1.2",
    "voyageai>=0.3.5",
    "openai>=1.35.13",
    "aiohttp>=3.9.5",
]

[project.scripts]
gss = "github-stars-search.main:cli"

# [tool.uv]
# package = true
